genai/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.0
sqlalchemy==2.0.23
aiofiles==23.2.1
pydantic==2.5.0
urlextract==1.8.0